fix: Exclude sample code from build
The sample code in the FTC project causes build failures with the downgraded dependency versions. This adds a config to the FTC project Gradle build that excludes those sample files from the built package. With that exclusion and the downgraded libraries the project builds cleanly, which should allow us to use the robot and drive hub as-is